benchmarks/competitive/live/requirements.txt
openai>=1.0
weaviate-client>=4.9
pymilvus>=2.5
qdrant-client>=1.12
chromadb>=0.6
httpx>=0.27
python-dotenv>=1.0
numpy>=1.26
tqdm>=4.66
psutil>=5.9
tabulate>=0.9
